requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==2.0.12
fake-useragent==1.1.3
idna==3.4
joblib==1.2.0
lxml==4.9.2
mpire==2.7.1
mysql-connector-python==8.0.33
protobuf==3.20.3
Pygments==2.15.1
requests==2.30.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==1.26.15